ContentView (Quantity)
Now have the quantity being displayed on the ContentView. The issue lay with explicitly stating that it was Int16 I was using both in the Extension of the model and in the NewEntryView.

Have changed the default amount of an item as zero is daft so the Stepper now starts at 1

+++ ShoppingItems/Core/ShoppingItems.swift
@@ -16,7 +16,7 @@ extension ShoppingItems {
         return NSFetchRequest<ShoppingItems>(entityName: "ShoppingItems")
     }
     @NSManaged public var itemToBeAdded: String
-    @NSManaged public var quantityOfItem: Int
+    @NSManaged public var quantityOfItem: Int16
     @NSManaged public var notesOnItem: String
     @NSManaged public var id: UUID
 }
